+++ src/ska_mid_cbf_tdc_mcs/subarray/subarray_component_manager.py
@@ -1921,9 +1921,10 @@ class CbfSubarrayComponentManager(CbfObsComponentManager):
                 status=TaskStatus.FAILED,
                 result=(
                     ResultCode.FAILED,
-                    "Failed to validate schema for Scan input JSON",
+                    "Failed to validate Scan input JSON",
                 ),
             )
+            return
 
         scan = json.loads(argin)