+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ This project adheres to `Semantic Versioning http://semver.org/>`_.
 UNRELEASED CHANGES
 ******************
 * SKB-803: Updated the artifacts PVC so that storageClassName and volumeMode can be configured from values
+* SKB-800: Subarray Scan command returns failure when receiving an invalid input JSON
 
 1.1.2
 *****